mummer: fix linux-64 build

+++ recipes/mummer/build.sh
@@ -48,10 +48,12 @@ for i in exact-tandems dnadiff mapview mummerplot nucmer promer run-mummer1 run-
 done
 
 for i in $binaries; do 
-  if file $MUMMER_HOME/$i | grep "Perl script"; then
-    # fix hashbang lines to use conda's perl
-    sed -i.bak '1 s|^#!/.*/perl$|#!/usr/bin/env perl|g' $MUMMER_HOME/$i
-    rm -rf $MUMMER_HOME/$i.bak
+  if [ "$(uname)" == "Darwin" ]; then
+    if file $MUMMER_HOME/$i | grep "Perl script"; then
+      # fix hashbang lines to use conda's perl
+      sed -i.bak '1 s|^#!/.*/perl$|#!/usr/bin/env perl|g' $MUMMER_HOME/$i
+      rm -rf $MUMMER_HOME/$i.bak
+    fi
   fi
   # ensure executable and setup symlink for binary
   chmod +x $MUMMER_HOME/$i
